trying, and failing, to get exact searching to work

+++ Fay_Thoma_Project1/src/Node.java
@@ -13,7 +13,7 @@ public class Node {
 	public Node(Interval i) {
 		interv = i;
 		//generate a random priority to maintain balanced treap
-		priority = new Random().nextInt(Integer.MAX_VALUE);
+		priority = new Random().nextInt(Short.MAX_VALUE);
 	}
 	
 	/**
